Implemented getProviders and Refined Conditions for deleteProvider and updateProvider Functions

+++ Backend/controllers/providers.js
@@ -1,5 +1,99 @@
 const Provider = require("../models/Provider");
 
+//@desc   Get all companies
+//@route  GET /api/v1/companies
+//@access Public
+exports.getProviders = async (req, res, next) => {
+    let query;
+
+    //Copy req.query
+    const reqQuery = {...req.query};
+
+    //Fields to exclude
+    const removeFields = ['select', 'sort', 'page', 'limit'];
+
+    //Loop over remove fields and delete them from reqQuery
+    removeFields.forEach(param => delete reqQuery[param]);
+    console.log(reqQuery);
+
+    //Create query string
+    let queryStr = JSON.stringify(reqQuery);
+    queryStr = queryStr.replace(/\b(gt|gte|lt|lte|in)\b/g, match => `$${match}`);
+
+    query = Provider.find(JSON.parse(queryStr))
+
+    //Select Fields
+    if (req.query.select) {
+        const fields = req.query.select.split(',').join(' ');
+        query = query.select(fields);
+    }
+
+    //Sort
+    if (req.query.sort) {
+        const sortBy = req.query.sort.split(',').join(' ');
+        query = query.sort(sortBy);
+    } else {
+        query = query.sort('name');
+    }
+    
+    //Pagination
+    const page = parseInt(req.query.page,10) || 1;
+    const limit = parseInt(req.query.limit,10) || 25;
+    const startIndex = (page - 1) * limit;
+    const endIndex = page * limit;
+    
+    try {
+        const total = await Provider.countDocuments();
+
+        query = query.skip(startIndex).limit(limit);
+
+        //Execute query
+        const providers = await query;
+
+        //Pagination result
+        const pagination = {};
+
+        if (endIndex < total) {
+            pagination.next = {
+                page: page + 1,
+                limit
+            }
+        }
+
+        if(startIndex > 0) {
+            pagination.prev = {
+                page: page - 1,
+                limit
+            }
+        }
+
+        res.status(200).json({
+            success: true, 
+            count: providers.length, 
+            data: providers
+        });
+    } catch (err) {
+        res.status(400).json({success: false});
+    }
+}
+
+//@desc   Get single company
+//@route  GET /api/v1/companies/:id
+//@access Public
+exports.getCompany = async (req, res, next) => {
+    try {
+        const company = await Company.findById(req.params.id);
+
+        if (!company) {
+            return res.status(400).json({succes: false});
+        }
+        res.status(200).json({
+            success: true,
+            data: company});
+    } catch (err) {
+        res.status(400).json({success: false});
+    }
+}
 //@desc Get peding providers
 //@route GET /api/v1/providers/pending
 //@access Private
@@ -63,14 +157,26 @@ exports.addProvider = async (req, res, next) => {
 //@access Private
 exports.updateProvider = async (req, res, next) => {
     try {
-        const provider = await Provider.findByIdAndUpdate(req.params.id, req.body, {
-            new: true,
-            runValidators: true
-        });
+
+        let provider = await Provider.findById(req.params.id)
+        
 
         if (!provider) {
             return res.status(400).json({success: false});
         }
+
+        if(req.user.role === 'user' || (provider.user.toString() !== req.user.id)) {
+            return res.status(401).json({
+              success:false,
+              message:`User ${req.user.id} is not authorized to update this provider's information`
+            })
+        }
+        
+        provider = await Provider.findByIdAndUpdate(req.params.id, req.body, {
+            new: true,
+            runValidators: true
+        });
+
 
         res.status(200).json({
             success: true,
@@ -90,6 +196,13 @@ exports.deleteProvider = async (req, res, next) => {
 
         if (!provider) {
             return res.status(400).json({success: false});
+        }
+
+        if(req.user.role === 'user' || (provider.user.toString() !== req.user.id)) {
+            return res.status(401).json({
+              success:false,
+              message:`User ${req.user.id} is not authorized to delete this provider's information`
+            })
         }
 
         await provider.deleteOne();
